add vpc nicspec
The why: the inescapable reality is that google_compute_instance
requires a list of nics and *not* a map. If we make a map, it's
hard to convert it. This is subtly wrong:

  [ module.vpc.subnet["a"], module.vpc.subnet["b"] ]

because it will fail a `tf destroy`: element "a" doesn't exist. So:

  [ try(module.vpc.subnet["a"], null), try(module.vpc.subnet["b"], null) ]

Cumbersome. While my change is making possbile to just use a ready-made list:

module.vpc.nicspec

Therefore I convert this module from "a random set of subnetworks",
to "objects dedicated to direct consumption by vmseries module(s)".