feat(vpc): networks_existing, merge output
Make the code equally focused on Terraform 0.12.x and 0.13.x.
It means `data { depends_on }` must be removed. For 0.12.29, the
consequences are quite dire as tested:

- Never run `terraform refresh`, it forces a subsequent
destroy/recreate of dependent resources. Recover from it by restoring
the tfstate from backup. No workaround available, which means that the
production environment can never be refreshed to fully re-read the
currently existing cloud resources.
- The `apply` will also destroy/recreate of multiple resources.
- For applying, always use: `terraform apply --refresh=false`
- For planning, always use: `terraform plan --refresh=false`

For outputs I use `merge()` which causes problems on 0.12.x in subsequent
for_each. But the workaround is available:
https://github.com/PaloAltoNetworks/terraform-best-practices/blob/d7865cd85fcf58b83c86cd816a12441e8aaec003/empty_state_and_tf12.md