fix(autoscale): vanishing file just after update
When init-cfg.txt is updated, the next `tf apply` first performs
Create and later Destroy. Thus the file is actually not present
in the bucket after the next `tf apply --refresh=true`.

It's the instance that needs the init-cfg.txt and so IGM should wait
here instead of the template.

The template shouldn't depend on that init-cfg.txt, moreover it passes
the create_before_destroy to the bucket causing the bug.