docs: change formatter from 'table' to 'document'
Most module inputs have very long descriptions, or very long default
values, or both. When formatting as a document instead of a table,
these descriptions more readable and prettier. Also the long unbreakable
strings, (such as URLs) do not force users to right-scroll the table
itself to see if the item is Required yes/no.

Additionally, hide the Providers section. As we generally require
a version of every provider with required_providers to avoid the
terraform warnings, all the providers tend to be listed in a directly
preceding section Requirements. It looks strange to repeat almost
exactly the same information. The default seems to be anticipating that
authors would not be using required_providers/version.